don't add/remove spaces to filenames
Rhino couldn't open any files. I looked into why it was having trouble; it was because
the filename was not what I had passed to zerocli. It seems like the reason for the
'remove the spaces' code was because extra spaces might get added inadvertently. So
this update is careful not to add the spaces and therefore doesn't have to remove them
later.